front/scripts/requirements.txt
# Python dependencies for hero icon extraction script
# Install with: pip install -r requirements.txt

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# OCR
pytesseract>=0.3.10
